Rewrote a lot of the TagCache class
This makes some pretty significant improvements in terms of usability and also efficiency. The mechanisms for adding and editing tag data have been greatly simplified and optimized for use with the serialization system. This should speed up tag conversion a bit because fewer I/O operations are needed to allocate tags now. In particular, new tags can now be reserved by calling the AllocateTag method, and then they can actually be given data later by using one of the two Overwrite methods. (Previously, the converter had to create a useless tag of size 1 and then overwrite it later...ew!)

However, this build may not be backwards-compatible with previous builds, as much of the TagCache interface has changed and some commands were tweaked. In particular, the "extract" and "import" commands will now always extract and import the header regardless of whether or not the "full" option is specified, and the "insert" command has been removed completely because tag data insertion no longer works in the same way. If there is enough demand for it, I can re-implement the "insert" command, but I don't believe that many people actually use it considering that the "tagblock" command is much more useful.